requirements.txt
asgiref==3.7.2
Brotli==1.1.0
dj-database-url==0.4.2
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
